Updated BD, Models and Controllers according to BD
